# Speed up metrics counter point submission